moving some files under `/checkout`
