+++ src/app/checkout/[show]/checkout-summary-card.tsx
@@ -3,7 +3,7 @@
 import { ChangeEvent } from 'react';
 
 import { Card } from '@/components';
-import { useCheckoutContext } from './checkout-context';
+import { useCheckoutContext } from '../checkout-context';
 
 interface CheckoutSummaryCardProps {
   price: number;
